remove multiple shapes
