+++ packages/joint-core/demo/org/src/org.js
@@ -1,4 +1,45 @@
-var graph = new joint.dia.Graph();
+const Member = joint.dia.Element.define('org.Member', {
+    size: { width: 180, height: 70 },
+    attrs: {
+        rect: { width: 170, height: 60 },
+
+        '.card': {
+            fill: '#FFFFFF', stroke: '#000000', 'stroke-width': 2,
+            'pointer-events': 'visiblePainted', rx: 10, ry: 10
+        },
+
+        image: {
+            width: 48, height: 48,
+            ref: '.card', 'ref-x': 10, 'ref-y': 5
+        },
+
+        '.rank': {
+            'text-decoration': 'underline',
+            ref: '.card', 'ref-x': 0.9, 'ref-y': 0.2,
+            'font-family': 'Courier New', 'font-size': 14,
+            'text-anchor': 'end'
+        },
+
+        '.name': {
+            'font-weight': '800',
+            ref: '.card', 'ref-x': 0.9, 'ref-y': 0.6,
+            'font-family': 'Courier New', 'font-size': 14,
+            'text-anchor': 'end'
+        }
+    }
+}, {
+    markup: '<g class="rotatable"><g class="scalable"><rect class="card"/><image/></g><text class="rank"/><text class="name"/></g>',
+});
+
+const Arrow = joint.dia.Link.define('org.Arrow', {
+    source: { selector: '.card' }, target: { selector: '.card' },
+    attrs: { '.connection': { stroke: '#585858', 'stroke-width': 3 }},
+    z: -1
+});
+
+const shapes = { ...joint.shapes, Member, Arrow };
+
+var graph = new joint.dia.Graph({}, { cellNamespace: shapes });
 
 var paper = new joint.dia.Paper({
     el: document.getElementById('paper'),
@@ -6,6 +47,7 @@ var paper = new joint.dia.Paper({
     height: 600,
     gridSize: 1,
     model: graph,
+    cellViewNamespace: shapes,
     perpendicularLinks: true,
     restrictTranslate: true
 });
