+++ packages/joint-core/src/shapes/index.mjs
@@ -1,11 +1,4 @@
-import * as basic from './basic.mjs';
 import * as standard from './standard.mjs';
 import * as devs from './devs.mjs';
-import * as chess from './chess.mjs';
-import * as erd from './erd.mjs';
-import * as fsa from './fsa.mjs';
-import * as org from './org.mjs';
-import * as pn from './pn.mjs';
-import * as uml from './uml.mjs';
 
-export { basic, standard, devs, chess, erd, fsa, org, pn, uml };
+export { standard, devs };
